requirements.txt
python-dotenv
langchain-huggingface
crewai
crewai-tools
